requirements.txt
anthropic>=0.40.0
tavily-python>=0.5.0
fastapi>=0.115.0
uvicorn>=0.32.0
sse-starlette>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
